+++ packages/PowerPack/src/tests/utils.ts
@@ -22,8 +22,8 @@ export const FUNC_TESTS: {[f: string]: {[test: string]: any}} = {
   Column: {'Column("age").length': 10},
   Value: {'Value("age", 0)': 61},
   RowCount: {'RowCount()': 10},
-  CumSum: {'CumSum(${age})': 61},
-  MovingAvg: {'MovingAvg(${age}, 1)': 61},
+  CumSum: {'CumSum(${age})': 61, 'CumSum(${age}, by=${sex})': 61, 'CumSum(${age}, orderBy=${subj}, descending=true)': 556},
+  MovingAvg: {'MovingAvg(${age}, 1)': 61, 'MovingAvg(${age}, 1, minPeriods=1)': 61},
   IndexOf: {'IndexOf(${subj}, ${subj})': 0},
   Lookup: {'Lookup("demog", "subj", ${subj}, "age")': 61},
 };